strangers fix

+++ media_backend/utils.js
@@ -2,7 +2,11 @@ const fs = require('fs');
 
 
 async function deleteImagesFromStorage(filenames) {
+    console.log("Deleting files: ", filenames)
+
     for (const filename in filenames) {
+        if (typeof filename !== 'string')
+            continue
         try {
             fs.unlinkSync(filename)
         } catch (e) {
